pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vocabmorizer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-login>=0.6",
    "flask-migrate>=4.0",
    "flask-wtf>=1.2",
    "flask-babel>=4.0",
    "sqlalchemy>=2.0",
    "authlib>=1.3",
    "bcrypt>=4.1",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
    "gunicorn>=22.0",
]

[project.optional-dependencies]
s3 = ["boto3>=1.34"]
dev = [
    "pytest>=8.0",
    "pytest-flask>=1.3",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
